Laser initialization

Each supported laser has its own (Julia) type which contains all the parameters required to decribe the electromagnetic field configuration. The simplest way to initialize a laser is via the setup_laser function.

LaserTypes.setup_laser — Method

setup_laser(laser, units; τ=nothing, kwargs...)

Initialize the specified laser using the default parameteres in the specified units. Supported units are

:SI: Values are in SI, the numeric types are regular numbers (Float64)

:SI_unitful: Values are in SI, but using Unitful.jl

:atomic: Values are in atomic units, the numeric types are regular numbers (Float64)

:atomic_unitful: Values are in atomic units, but using UnitfulAtomic.jl

The keyword arguments can be used to give specific values to the parameters instead of using the defaults. You can specify parameteres such as the wavelength and beam waist via λ and w₀. The duration of the pulse (assuming Gaussian temporal profile) can be given via τ. Be sure to use the same units as the ones provided via units. You can also specify dimensionless parameteres such as the normalized vector potential (a₀), the initial phase (ϕ₀) and the polarization (ξx and ξy). See the docomuntation for each laser type for more details on the supported arguments.