Advanced usage

If you need more control over the parameters, you can use the constructors for the laser types directly.

LaserTypes.GaussLaser — Type

struct GaussLaser{V,Q,M,L,F,C,T,P,W,K,E}

The GaussLaser is defined by the following independent parameters

c is the speed of light in vaccum, with the default value being in SI (c_0 from the CODATA2018 in the PhysicalConstants package)

q is the electric charge, with the default value being the one for the electron in SI (-e from the CODATA2018 in the PhysicalConstants package)

m_q is the mass of the charge, with the default value being the one for the electron in SI (m_e from the CODATA2018 in the PhysicalConstants package)

λ is the laser wavelangth with the default value 0.8μm

a₀ is the normalized vector potential (defined as $a_0=\frac{eA}{m_e c^2}$)

ϕ₀ is the initial phase with the default value 0.0

w₀ is the beam radius at the Rayleigh range or beam waist with the default value 58.0μm

ξx and ξy give the polarization and have the default value 1.0 + 0im and 0.0 + 0im

profile is the temporal profile of the pulse and the default one is a Gaussian one

During the initialization of the GaussLaser type, some useful derived values are also computed

ω is the angular frequency and is given by $2π c / λ$

k is the wavenumber and is given by $2π / λ$

z_R is the Rayleigh range and is given by $k w_0^2 / 2$

T₀ is the laser period and is given by $2π / ω$

E₀ is the amplitude of the electric field and is given by $a_0\frac{m_q c \omega}{q}$

LaserTypes.LaguerreGaussLaser — Type

struct LaguerreGaussLaser{V,Q,M,L,F,C,T,P,I,W,K,E,R}

The LaguerreGaussLaser is defined by the following independent parameters

c is the speed of light in vaccum, with the default value being in SI (c_0 from the CODATA2018 in the PhysicalConstants package)

q is the electric charge, with the default value being the one for the electron in SI (-e from the CODATA2018 in the PhysicalConstants package)

m_q is the mass of the charge, with the default value being the one for the electron in SI (m_e from the CODATA2018 in the PhysicalConstants package)

λ is the laser wavelangth with the default value 0.8μm

a₀ is the normalized vector potential (defined as $a_0=\frac{eA}{m_e c^2}$)

ϕ₀ is the initial phase with the default value 0.0

w₀ is the beam radius at the Rayleigh range or beam waist with the default value 58.0μm

ξx and ξy give the polarization and have the default value 1.0 + 0im and 0.0 + 0im

profile is the temporal profile of the pulse and the default one is a Gaussian one

p is the radial index of the mode, $p ∈ ℤ, p ≥ 0$, with the default value 1

m is the azimuthal index of the mode, $m ∈ ℤ$, with the default value 0

During the initialization of the LaguerreGaussLaser type, some useful derived values are also computed

ω is the angular frequency and is given by $2π c / λ$

k is the wavenumber and is given by $2π / λ$

z_R is the Rayleigh range and is given by $k w_0^2 / 2$

T₀ is the laser period and is given by $2π / ω$

E₀ is the amplitude of the electric field and is given by $a_0\frac{m_q c \omega}{q}$

Nₚₘ is a normalization factor given by $\sqrt{(p+1)_{|m|}}$, with $(x)_n$ the Pochhammer symbol

Note

When specifying derived values, the corresponding independent parameters must also be given in order to avoid discrepancies. For example if you give k, you must also give the corresponding λ.