Temporal Profiles

The complete expression of the electromaginetic fields describing the laser pulse is given by the combination of the spatial part together with the time dependence.

LaserTypes.g — Method

g(z, t, par)

The time dependence of the fields defined by this package is given by

\[g(z, t) = \exp(ω t) envelope(z, t),\]

where

z and t are the position on the $Oz$ axis and the time

par are the laser parameters which pass the corresponding profile to the envelope

and

$ω$ is the angular frequency of the laser pulse

$envelope(z, t)$ is a function that can be used to control the duration of the pulse

There are several types that provide profiles to the temporal envelope:

LaserTypes.ConstantProfile — Type

struct ConstantProfile

This is the trivial profile

\[envelope(z, t) = 1\]

which gives an infinite duration for the electromagnetic field configuration (we cannot call this a laser pusle since a pulse implicitely has a finite duration).

LaserTypes.Cos²Profile — Type

Cos²Profile{V,T,L}

This envelope provides a finite duration for the laser pulse and thus can provide a more realistic description of an actual laser pulse.

\[envelope(z, t) = \begin{cases} \cos\left[\left(\frac{φ}{τ}\right)\right]^2, & \text{for } |t - t₀| < τ / 2\\ 0 \,, & \text{otherwise} \end{cases}\]

where

\[\varphi = (t - t_0) - \frac{z - z_0}{c}\,,\]

and

c is the speed of light

τ is the duration of the pulse and has the default value 18.02fs

t₀ is the origin of the time axis and it is 0 by default

z₀ is the initial position of the intensity peak and has the default value -4*τ*c

LaserTypes.GaussProfile — Type

GaussProfile{V,T,L}

This envelope provides a finite duration for the laser pulse and thus can provide a more realistic description of an actual laser pulse.

\[envelope(z, t) = \exp\left[\left(\frac{φ}{τ}\right)^2\right],\]

where

\[\varphi = (t - t_0) - \frac{z - z_0}{c}\,,\]

and

c is the speed of light

τ is the duration of the pulse (FWHM) and has the default value 18.02fs

t₀ is the origin of the time axis and it is 0 by default

z₀ is the initial position of the intensity peak and has the default value -4*τ*c

LaserTypes.QuasiRectangularProfile — Type

QuasiRectangularProfile{V,T,L}

This envelope produces a pulse with a predominant constant part of width $Δz$ which could offer better results than the Gaussian profile in the paraxial limit (which is considered for the spatial profiles). The shape of the envelope is given by

\[envelope(z, t) = \begin{cases} \exp\left[\left(\frac{φ}{τ}\right)^2\right], & \text{for } φ ≤ 0\\ 1\,, & \text{for } φ < Δz\\ \exp\left[\left(\frac{φ - Δz/c}{τ}\right)^2\right], & \text{otherwise} \end{cases}\]

where

\[\varphi = (t - t_0) - \frac{z - z_0}{c}\,,\]

and

c is the speed of light

τ is the duration of the pulse (FWHM) and has the default value 18.02fs

t₀ is the origin of the time axis and it is 0 by default

z₀ is the initial position of the intensity peak and has the default value -4*τ*c

Δz is the width of the flat part of the profile and the default value 10*τ*c